# openapi.yaml
openapi: 3.1.0
info:
  title: ForemanOS API
  version: 1.0.0
  summary: AI-powered construction project management API
  description: |
    AI-powered construction project management platform API.

    **Authentication:** Most endpoints require JWT authentication via NextAuth.

    **Rate Limits:**
    - Chat API: 20 requests/minute
    - Upload API: 10 requests/minute
    - Auth API: 5 requests/5 minutes
    - General API: 60 requests/minute

    **Subscription Tiers:** Free, Starter, Pro, Team, Business, Enterprise
  contact:
    name: ForemanOS Support
    url: https://foremanos.site
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://foremanos.site
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /api/stripe/webhook:
    post:
      tags:
        - Stripe
      summary: Stripe webhook handler
      description: |
        Webhook endpoint for Stripe events. Handles subscription lifecycle events.

        **Handled Events:**
        - checkout.session.completed
        - customer.subscription.created
        - customer.subscription.updated
        - customer.subscription.deleted
        - invoice.paid
        - invoice.payment_failed

        **Security:** Requires valid Stripe signature header
      operationId: stripeWebhook
      security: []
      parameters:
        - name: stripe-signature
          in: header
          required: true
          schema:
            type: string
          description: Stripe webhook signature for verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Stripe event payload (varies by event type)
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                    example: true
                  duplicate:
                    type: boolean
                    description: Whether this was a duplicate event
        '400':
          description: Invalid signature or payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Webhook Error: Invalid signature"
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth JWT token from session
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Internal server error"
              details:
                type: string
                description: Additional error details (dev mode only)
...